+++ src/Symfony/Component/Routing/Tests/Fixtures/AnnotationFixtures/PrefixedActionPathController.php
@@ -2,7 +2,7 @@
 
 namespace Symfony\Component\Routing\Tests\Fixtures\AnnotationFixtures;
 
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
 /**
  * @Route("/prefix", host="frankdejonge.nl", condition="lol=fun")