+++ src/Symfony/Component/Serializer/Tests/Annotation/DiscriminatorMapTest.php
@@ -12,7 +12,7 @@
 namespace Symfony\Component\Serializer\Tests\Annotation;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Serializer\Annotation\DiscriminatorMap;
+use Symfony\Component\Serializer\Attribute\DiscriminatorMap;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 
 /**