+++ src/Symfony/Component/Serializer/Tests/Fixtures/Annotations/GroupDummy.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Serializer\Tests\Fixtures\Annotations;
 
-use Symfony\Component\Serializer\Annotation\Groups;
+use Symfony\Component\Serializer\Attribute\Groups;
 use Symfony\Component\Serializer\Tests\Fixtures\ChildOfGroupsAnnotationDummy;
 use Symfony\Component\Serializer\Tests\Fixtures\Annotations\GroupDummyInterface;