+++ src/Symfony/Component/Serializer/Tests/Fixtures/Annotations/SerializedNameDummy.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Serializer\Tests\Fixtures\Annotations;
 
-use Symfony\Component\Serializer\Annotation\SerializedName;
+use Symfony\Component\Serializer\Attribute\SerializedName;
 
 /**
  * @author Fabien Bourigault <bourigaultfabien@gmail.com>