+++ src/Symfony/Component/Serializer/Tests/Fixtures/Annotations/SerializedPathDummy.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Serializer\Tests\Fixtures\Annotations;
 
-use Symfony\Component\Serializer\Annotation\SerializedPath;
+use Symfony\Component\Serializer\Attribute\SerializedPath;
 
 /**
  * @author Tobias Bönner <tobi@boenner.family>